Moved into cpp
